feat[#303] : 비밀번호 확인에서 컨펌 모달 제거
