+++ src/features/setting/hooks/index.ts
@@ -1,3 +1,2 @@
-export * from "./usePasswordCheckModal.tsx";
 export * from "./usePasswordChangeModal.tsx";
 export * from "./useSettingPermission.tsx";
